kubevirt: make kubemacpool optional
Don't install kubemacpool by default, as it recently causes issues and
we don't use it.

The error we are facing is:
```
Internal error occurred: failed calling webhook "mutatevirtualmachines.kubemacpool.io": failed to call webhook: Post "https://kubemacpool-service.kubemacpool-system.svc:443/mutate-virtualmachines?timeout=10s": proxy error from 127.0.0.1:6443 while dialing 10.42.0.39:8000, code 502: 502 Bad Gateway
```
This is due to the controller manager crashing because it wants to
recreate the config map and errors out when it detects that the map
already exists.

If `install_kubemacpool` variable is defined install it anyway.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>